Sistema de tiro com mouse

+++ src/entities/BulletShoot.java
@@ -9,12 +9,12 @@ import world.Camera;
 
 public class BulletShoot extends Entity {
 	
-	private int dx;
-	private int dy;
+	private double dx;
+	private double dy;
 	private double speed = 4;
 	private int life = 60, curLife = 0;
 	
-	public BulletShoot(double x, double y, int width, int height, BufferedImage sprite, int dx, int dy) {
+	public BulletShoot(double x, double y, int width, int height, BufferedImage sprite, double dx, double dy) {
 		super(x, y, width, height, sprite);
 		this.dx = dx;
 		this.dy = dy;
